+++ src/main/java/ru/kata/spring/boot_security/demo/repository/RoleRepository.java
@@ -5,7 +5,6 @@ import org.springframework.stereotype.Repository;
 import ru.kata.spring.boot_security.demo.models.Role;
 
 @Repository
-public interface RoleRepository extends JpaRepository<Role, Integer> {
-
-    Role findRoleById(Integer id);
+public interface RoleRepository extends JpaRepository<Role, Long> {
+    Role findRoleById(Long id);
 }
