+++ src/main/java/ru/kata/spring/boot_security/demo/repository/UserRepository.java
@@ -7,8 +7,8 @@ import ru.kata.spring.boot_security.demo.models.User;
 
 
 @Repository
-public interface UserRepository extends JpaRepository <User, Integer> {
+public interface UserRepository extends JpaRepository <User, Long> {
     User findByEmail(String email);
 
-    User findUserById(Integer id);
+    User findUserById(Long id);
 }
